Multiple sinker model: add option to load inclusion origins from file
For model 7 (pTatin sinkers, 3D only), add a command line option to specify a
text file with coordinates for the inclusion centers, as three floating
point numbers per line.

Note that no checks are done to ensure that these inclusions are in the
domain, don't overlap with specified radius, etc.